Creates css and js files and sets up basic folder structure for views
The main overall layout as well as the main content view have both been created. At this point
the plan for views is to use one view and pull in partials conditionally when the appropriate
data from an ajax request is received.

A plan for the initial positioning of elements via Javascript is also formed.