On branch reverse   Changes to be committed:  	modified:   public/css/main.css  	modified:   public/js/draw.js  	modified:   public/js/main.js
Create a new branch to flesh out the possibility of aligning everything from the index node out dynamically
around a circle or bezier curve with GSAP bezier plugin.

Added functions for splitting the nodes array and for finding the paths the nodes will be aligned on.
The latter will remain incomplete until the dimensions of the index node's circle are determined, points along it calculated, and the math for creating the outer arcs from it settled on.