On branch reverse  Changes to be committed: 	modified:   public/js/draw.js
Position the nodes along a Raphael svg path and segments of the path so each node animated along the same path, but stopped at different places along it. Used Raphael to get values for the GSAP's bezier plugin.

This function will likely be moved to the main js file and split into two functions that manage the animation and the initial/final positioning respectively.