On branch reverse  Changes to be committed: 	modified:   public/js/draw.js
Start function that creates crossing paths--one on the node and one between the center of the node and the center of the index. The line on the node is determined by which quadrant the node is in. The intersection point, determined with Raphael, will serve as the start point of the svg between be index and the node.