On branch reverse  Changes to be committed: 	modified:   public/js/draw.js   modified:   public/js/main.js
Add functions to determine the quadrant a given node is in, the corners of the node on the side facing the index node, and the point where a line from the center of the index to the center of the node would cross. These together get the starting point for the svg paths connecting the the node to the index. The next step is to draw the svg connection for each node.